organize into edit_mode folder
